Added download/upload photos for pet, made Response with codes in controller's returns

+++ src/main/java/ru/pethelper/service/impl/UserServiceImpl.java
@@ -168,7 +168,7 @@ public class UserServiceImpl implements UserService, UserDetailsService {
 
     @Override
     public void saveImage(long userId, MultipartFile image) throws Exception {
-        UserEntity user = userRepo.getOne((long) userId);
+        UserEntity user = userRepo.getOne(userId);
         user.setUserImage(image.getBytes());
         userRepo.save(user);
     }
